agent/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
Authlib==1.6.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
cssselect2==0.8.0
Deprecated==1.2.18
dnspython==2.7.0
docstring_parser==0.16
fastapi==0.115.12
fonttools==4.58.1
google-adk==1.1.1
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.95.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.17.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.2
gridfs==0.0.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.6.1
incremental==22.10.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mcp==1.9.2
mdurl==0.1.2
numpy==2.2.6
oauthlib==3.2.2
opentelemetry-api==1.33.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.17.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pymongo==4.13.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyphen==0.17.2
pyrsistent==0.20.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
rich==13.7.1
rsa==4.9.1
service-identity==24.1.0
setuptools==68.1.2
shapely==2.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.5
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.34.2
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
wheel==0.42.0
wrapt==1.17.2
zipp==3.22.0
zope.interface==6.1
zopfli==0.2.3.post1
